Refactor : DeleteStockModal
